src/main/java/name/guolanren/_1to100/_11to20/p11/ContainerWithMostWater.java
package name.guolanren._1to100._11to20.p11;

/**
 * @link https://leetcode.com/problems/container-with-most-water/
 * @author guolanren
 */
public class ContainerWithMostWater {

    public int maxArea(int[] height) {
        int left = 0;
        int right = height.length - 1;
        int max = 0;

        while (left < right) {
            max = Math.max(max, (right - left) * Math.min(height[left], height[right]));
            if (height[left] >= height[right]) {
                right--;
            } else {
                left++;
            }
        }
        return max;
    }

}
